Remove default class attributes for now
Adding these helpers made it easier to get some autocompletion when
creating a Service in an IDE, but made exceptions less helpful because
it started trying to do string interpolation with Nones instead of
loudly saying that a Service didn't have a domain defined. In the future
it might still be valuable to do this, raising a more useful exception
explicitly if a domain, service_name, or host_resolver_class is not
specified by the user.

on-behalf-of: @ithaka <dane.hillard@ithaka.org>