Improve inheritance structure (#49)
* Improve inheritance structure

Before this change, `Service` and `DiscoverableService` were tightly
coupled, and `DiscoverableService` overrode most of the behavior
explicitly defined in `Service`. Instead it makes more sense that these
both inherit from some shared ancestor, and specialize only what
differs. This allowed `ServiceMeta` to know less about the classes that
derive from it as well.

* Remove default class attributes for now

Adding these helpers made it easier to get some autocompletion when
creating a Service in an IDE, but made exceptions less helpful because
it started trying to do string interpolation with Nones instead of
loudly saying that a Service didn't have a domain defined. In the future
it might still be valuable to do this, raising a more useful exception
explicitly if a domain, service_name, or host_resolver_class is not
specified by the user.

* Update CHANGELOG with information about inheritance change