Forward the `json` kwarg to requests
In addition to `data`, requests accepts a `json` kwarg that allows you
to pass in a `dict` that will eventually be serialized into a JSON POST
body (as compared to `data`, which gets form-encoded first). This was an
oversight in the initial implementation since that wasn't a use for us
at the time.

on-behalf-of: @ithaka <dane.hillard@ithaka.org>