Add linting stage to Travis CI build (#55)
* Add linting stage to Travis CI build

Makes use of the concept of "stages" on Travis CI. There is one stage
each here for linting, testing, and deploying. The deploy stage is now
cut off at the pass if the target is not a tag from the master branch,
so the build should be slightly faster.

* Reorder Pythons in Travis CI matrix and target 3.4 with black

Stages get the first Python listed by default, but black can only run in
Python 3.6+. Also make sure that black checks against 3.4-compatible
syntax (black should also be run with this flag when transforming
files).

* Process code using black

Reformat the code using black to establish the baseline. Used `black -t
py34 apiron tests` to produce the change.

* Install black in lint stage only

black requires Python 3.6+ and won't even install in lower versions.
Remove black from dev-requirements.txt and only install it manually in
the lint stage where Python 3.6+ is guaranteed.

* Document .travis.yml sections

Add comments to .travis.yml for clarity.

* Add note about black to CONTRIBUTING.md

* Add pyflakes checks to linting stage

Add `pyflakes` to linting stage and fix import warnings it was raising.
I had previously said `import *` was fine in apiron/__init__.py, but
pyflakes does get mad about that. Explode that out to the items that
really are there, and list them out explicitly in `__all__`.

* Blackify apiron/__init__.py

* Update black configuration

* Use `pyproject.toml` to configure black for all supported Python
versions; this reduces likelihood of running black the wrong way
* Increase line length to 120 for consistency with existing code style