Use descriptor to simplify service calling
Using the descriptor protocol allows for inspecting the service that
owns an endpoint, giving an easier entrypoint for applying
ServiceCaller.call than doing this in the ServiceMeta metaclass.

I had wanted exactly this in the beginning but didn't make the
connection that descriptorss would get me there. It feels like this
might improve the road to better code introspection as laid out in
ithaka/apiron#54.

on-behalf-of: @ithaka <dane.hillard@ithaka.org>