Fix condition for deployments from Travis CI
Per Travis CI, because a tag is never associated with a branch,
receiving a webhook caused by a tag won't ever meet the condition that
the branch is master. Additionally, since the criteria for the deploy
stage is specified in the stage configuration, it no longer needs to be
in the provider configuration.

on-behalf-of: @ithaka <dane.hillard@ithaka.org>