Remove ServiceCaller and path_kwargs vestiges (#66)
* Remove ServiceCaller

The `ServiceCaller` class is not needed; all methods inside it were
static or class methods that only needed the class reference in order to
call its other methods. Moving these to the module level and updating
references keeps all functionality working, and will mean there is only
one right way of using apiron (the `SomeService.some_endpoint(...)`
approach).

on-behalf-of: @ithaka <dane.hillard@ithaka.org>

* Remove path_kwargs

The `path_kwargs` argument was deprecated in a previous version in favor
of passing path placeholder fillers as additional keyword arguments to
`call`. In favor of one right way of doing things, this old way is now
removed.

on-behalf-of: @ithaka <dane.hillard@ithaka.org>

* Remove unused import

on-behalf-of: @ithaka <dane.hillard@ithaka.org>

* Blacken test module

on-behalf-of: @ithaka <dane.hillard@ithaka.org>

* Add upgrade guide and brush up docs

An upgrade guide is important for anyone using 2.X who wants to smoothly
transition to 3.X. If in the future we make other breaking changes, it
can be added to/updated to reflect the changes needed for 4.X and
beyond.

on-behalf-of: @ithaka <dane.hillard@ithaka.org>

* Add removed path_kwargs docstring for **kwargs

on-behalf-of: @ithaka <dane.hillard@ithaka.org>