Fix building of docs on readthedocs.io (#81)
Because the management of dependencies for docs, linting, and so on
moved into setup.cfg, this has to be accounted for in order for
ReadTheDocs to know how to build our documentation. The process can be
customized by adding this configuration file which takes the place of
the online configuration.