Bump pytest to more recent version
The Travis CI build is currently failing. Per [this StackOverflow
answer](https://stackoverflow.com/a/58189684/2077393), it's because
pytest's dependency on attrs got a bit funky. Updating to a more recent
pytest version should hopefully fix this.

For now, I'm bumping this to the last version of pytest that supports
Python 3.4; going to pytest 5+ would mean having the discussion about
when to drop support for Python 3.4, which is good to have soon but may
not be strictly required to fix this.

on-behalf-of: @ithaka <dane.hillard@ithaka.org>