Remove legacy Python version support
Although it is important to maintain support for an older Python version
with a demonstrated active user base, it's important to assess the
landscape occasionally to make sure a package isn't languishing
needlessly for what are only percieved support requirements.

Per [PyPI Stats](https://pypistats.org/packages/apiron) there are no
regular users of `apiron` on Pythons 3.4 or 3.5. By removing support for
these versions, we're able to benefit from new features like f-strings
and type hinting along with newer versions of peripheral tooling like
`pytest`.

Although this package doesn't have a particularly pressing need to
upgrade, it's good to get into the practice of this kind of refresh.

on-behalf-of: @ithaka <dane.hillard@ithaka.org>