Include tests in shipped package
GNU Health is using apiron for their Orthanc integration, and as part of
that effort the folks working on openSUSE [would like to include
apiron](https://build.opensuse.org/request/show/752387)
in their package manager repository so that it can be easily installed
by their users.

The tests aren't currently included in the package we publish (an
intentional decision originally made
[here](https://github.com/ithaka/apiron/commit/2053ab2c11ff86f3a973844e1abf2811421dcbe6)),
but they would act as a convenient verification step for the package
after it's installed by the package manager.

on-behalf-of: @ithaka <dane.hillard@ithaka.org>